Cluster grenades burn things for 4 seconds
Things can now burn for different lengths of time, depending on the
ignition weapon. This is done with a new field in FireMixin that
checks and stores the name of the weapon that started the fire.
Currently, only the cluster grenade and flamethrower have
dedicated burn times; anything burned by another source will burn
for the same time as if it were burned by the flamethrower. (Cluster
grenades and flamethrowers are the only sources of fire currently.)